requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.7
django-admin-tools==0.9.3
django-axes==7.0.2
django-cachalot==2.7.0
django-ckeditor==6.7.2
django-clearcache==1.2.1
django-environ==0.12.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-redis==5.4.0
django-rest-swagger==2.2.0
django-secure==1.0.2
djangorestframework==3.15.2
drf-yasg==1.21.10
Faker==37.0.0
feedparser==6.0.11
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.2
mypy==1.15.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sgmllib3k==1.0.0
simplejson==3.20.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
